ui: extract security page cards into components and fix sidebar hover
- Extract PasswordCard, PasskeyCard, DangerZoneCard from Security page
- Security.tsx reduced from 207 to 44 lines
- Fix sign out button hover using solid primary-bg instead of subtle overlay
- Name MFA visibility conditions as consts for readability
